Move `Event` to standalone class
